requirements.txt
asgiref==3.8.1
attrs==23.2.0
base-62==0.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
djangorestframework-camel-case==1.4.2
djangorestframework-recursive==0.1.2
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
sqlparse==0.5.0
swagger-spec-validator==3.0.3
typing_extensions==4.11.0
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.1
user-agents==2.2.0
